# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DT Translation API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/translations/export": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Export translations",
                "description": "Exports translations based on the given locale and optional tag.",
                "operationId": "9f227c61c766f0f5cacdb54bab6df4a2",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "query",
                        "description": "Locale code for translations",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "en"
                        }
                    },
                    {
                        "name": "tag",
                        "in": "query",
                        "description": "Optional tag to filter translations",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "mobile"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Translations exported successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Translations exported successfully."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "key": {
                                                        "type": "string",
                                                        "example": "in-quae-amet-sequi-possimus-inventore-expedita"
                                                    },
                                                    "content": {
                                                        "type": "string",
                                                        "example": "Et voluptatem rerum accusantium quis aut illum ut dolorem."
                                                    },
                                                    "locale": {
                                                        "type": "string",
                                                        "example": "en"
                                                    },
                                                    "tags": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                },
                                                                "name": {
                                                                    "type": "string",
                                                                    "example": "mobile"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}